+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix clojure-lsp lint crash when analyzing specific macros with clj-kondo hooks.
+
 ## 2021.03.01-19.18.54
 
 - Bump clj-kondo fixing issues on require form not being analyzed.