+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2021.03.05-13.35.47
+
 - Fix clojure-lsp lint crash when analyzing specific macros with clj-kondo hooks.
 - Fix didChange for Nvim client.
 - Add new clojure-lsp linter: unused-public-var - Fixes #359